factor out versioned cipher code into dedicated module
